[FEAT] #11 - Database insert return type 변경 Int > Long

+++ core/datasource/src/main/java/com/core/datasource/PostDatasource.kt
@@ -9,7 +9,7 @@ import kotlinx.coroutines.flow.Flow
  */
 interface PostDatasource {
     // post 추가
-    suspend fun addPost(post: Post): Int
+    suspend fun addPost(post: Post): Long
 
     // 특정 년도/월/일 post 요청
     suspend fun getPost(year: Int, month: Int, day: Int): Post?
