+++ core/datasource/src/main/java/com/core/datasource/model/Image.kt
@@ -6,11 +6,11 @@ import com.core.database.model.ImageEntity
  * datasource Image 정보
  */
 data class Image(
-    val id: Int?, // 이미지 id
+    val id: Long?, // 이미지 id
     val imageUrl: String // 이지미 url
 )
 
-internal fun Image.toImageEntity(postId: Int) = ImageEntity(
+internal fun Image.toImageEntity(postId: Long) = ImageEntity(
     id = id,
     postId = postId,
     imageUrl = imageUrl
